add Logger class

+++ lib/Utils.hpp
@@ -12,6 +12,7 @@
 #include <algorithm>
 #include <cmath>
 #include <limits>
+#include "Logger.hpp"
 
 namespace Utils {
 namespace Counter {
@@ -54,7 +55,7 @@ inline void LIB_THROW(const char* message) { throw LibException(message); }
 }
 
 namespace Consts {
-constexpr double NAN_INT = std::numeric_limits<int>::quiet_NaN();
+constexpr int NAN_INT = std::numeric_limits<int>::quiet_NaN();
 constexpr double NAN_DOUBLE = std::numeric_limits<double>::quiet_NaN();
 constexpr double POS_INF_DOUBLE = std::numeric_limits<double>::infinity();
 constexpr double NEG_INF_DOUBLE = -std::numeric_limits<double>::infinity();
